# api/openapi.yaml
openapi: 3.0.4
info:
  title: UnderSound Content API
  description: |
    API de contenido inspirada en Bandcamp para gestionar **álbumes, pistas, merchandising, labels y comentarios**.
    Incluye paginación, filtros, subida de archivos (multipart), acciones de publicación y favoritos.
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/escuelapolitecnicaca/contenido/1.0.0
  description: SwaggerHub API Auto Mocking
- url: https://api.undersound.example.com/v1
  description: Producción
- url: http://localhost:8080
  description: Local
security:
- oauth2:
  - content.read
  - content.write
...
paths:
  /albums:
    get:
      tags:
      - Albums
      summary: Listar álbumes
      operationId: albumsGET
      parameters:
      - name: page
        in: query
        required: false
        style: form
        explode: true
        schema:
          minimum: 1
          type: integer
          default: 1
      - name: limit
        in: query
        required: false
        style: form
        explode: true
        schema:
          maximum: 100
          minimum: 1
          type: integer
          default: 20
      - name: include
        in: query
        description: "Campos relacionados a incluir, separados por coma. Ej. `tracks,label,stats`"
        required: false
        style: form
        explode: false
        schema:
          type: array
          items:
            type: string
      - name: artistId
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          format: uuid
      - name: labelId
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          format: uuid
      - name: genre
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: tag
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: releaseState
        in: query
        description: Estado de publicación
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - draft
          - scheduled
          - published
          - archived
      - name: q
        in: query
        description: Búsqueda full-text por título/descripcion
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: Lista paginada de álbumes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAlbumList"
      x-swagger-router-controller: Albums
    post:
      tags:
      - Albums
      summary: Crear álbum
      operationId: albumsPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlbumCreateInput"
        required: true
      responses:
        "201":
          description: Álbum creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlbumResponse"
        "400":
          description: Petición inválida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-swagger-router-controller: Albums
  /albums/{albumId}:
    get:
      tags:
      - Albums
      summary: Obtener detalle de un álbum
      operationId: albumsAlbumIdGET
      parameters:
      - name: albumId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      - name: include
        in: query
        description: "Campos relacionados a incluir, separados por coma. Ej. `tracks,label,stats`"
        required: false
        style: form
        explode: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: Detalle de álbum
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlbumResponse"
        "404":
          description: Recurso no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-swagger-router-controller: Albums
    delete:
      tags:
      - Albums
      summary: Eliminar o archivar lógicamente un álbum
      operationId: albumsAlbumIdDELETE
      parameters:
      - name: albumId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      - name: hard
        in: query
        description: "Si true, borrado físico irreversible"
        required: false
        style: form
        explode: true
        schema:
          type: boolean
          default: false
      responses:
        "204":
          description: Eliminado
      x-swagger-router-controller: Albums
    patch:
      tags:
      - Albums
      summary: Actualizar parcialmente un álbum
      operationId: albumsAlbumIdPATCH
      parameters:
      - name: albumId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlbumUpdateInput"
        required: true
      responses:
        "200":
          description: Álbum actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlbumResponse"
      x-swagger-router-controller: Albums
  /albums/{albumId}/publish:
    post:
      tags:
      - Albums
      summary: Publicar álbum
      operationId: albumsAlbumIdPublishPOST
      parameters:
      - name: albumId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Publicado. Retorna recurso actualizado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlbumResponse"
      x-swagger-router-controller: Albums
  /albums/{albumId}/tracks:
    post:
      tags:
      - Albums
      summary: Añadir pistas a un álbum
      operationId: albumsAlbumIdTracksPOST
      parameters:
      - name: albumId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/albumId_tracks_body"
        required: true
      responses:
        "200":
          description: Álbum con pistas actualizadas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlbumResponse"
      x-swagger-router-controller: Albums
  /albums/{albumId}/tracks/{trackId}:
    delete:
      tags:
      - Albums
      summary: Quitar una pista del álbum
      operationId: albumsAlbumIdTracksTrackIdDELETE
      parameters:
      - name: albumId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      - name: trackId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: Eliminado
      x-swagger-router-controller: Albums
  /albums/{albumId}/cover:
    post:
      tags:
      - Uploads
      - Albums
      summary: Subir o actualizar portada del álbum
      operationId: albumsAlbumIdCoverPOST
      parameters:
      - name: albumId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/albumId_cover_body"
        required: true
      responses:
        "200":
          description: Álbum con portada actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlbumResponse"
      x-swagger-router-controller: Uploads
  /albums/{albumId}/comments:
    get:
      tags:
      - Comments
      - Albums
      summary: Listar comentarios del álbum
      operationId: albumsAlbumIdCommentsGET
      parameters:
      - name: albumId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      - name: page
        in: query
        required: false
        style: form
        explode: true
        schema:
          minimum: 1
          type: integer
          default: 1
      - name: limit
        in: query
        required: false
        style: form
        explode: true
        schema:
          maximum: 100
          minimum: 1
          type: integer
          default: 20
      responses:
        "200":
          description: Lista paginada de comentarios
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCommentList"
      x-swagger-router-controller: Comments
    post:
      tags:
      - Comments
      - Albums
      summary: Comentar en un álbum
      operationId: albumsAlbumIdCommentsPOST
      parameters:
      - name: albumId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentCreateInput"
        required: true
      responses:
        "201":
          description: Comentario creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentResponse"
        "403":
          description: "No autorizado para comentar (por ejemplo, solo si ha comprado)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-swagger-router-controller: Comments
  /albums/{albumId}/stats:
    get:
      tags:
      - Albums
      summary: Estadísticas de un álbum
      operationId: albumsAlbumIdStatsGET
      parameters:
      - name: albumId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      - name: range
        in: query
        description: Rango temporal para agregados
        required: false
        style: form
        explode: true
        schema:
          type: string
          default: all
          enum:
          - 7d
          - 30d
          - all
      responses:
        "200":
          description: Estadísticas agregadas del álbum
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlbumStats"
      x-swagger-router-controller: Albums
...
  /labels/{labelId}/albums:
    get:
      tags:
      - Labels
      - Albums
      summary: Discografía de la label
      operationId: labelsLabelIdAlbumsGET
      parameters:
      - name: labelId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      - name: page
        in: query
        required: false
        style: form
        explode: true
        schema:
          minimum: 1
          type: integer
          default: 1
      - name: limit
        in: query
        required: false
        style: form
        explode: true
        schema:
          maximum: 100
          minimum: 1
          type: integer
          default: 20
      responses:
        "200":
          description: Lista paginada de álbumes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAlbumList"
      x-swagger-router-controller: Labels
...
components:
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
        details:
          type: object
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 123
      example:
        total: 123
        limit: 20
        page: 1
    Image:
      type: object
      properties:
        url:
          type: string
          format: uri
        alt:
          type: string
        width:
          type: integer
        height:
          type: integer
      example:
        alt: alt
        width: 6
        url: http://example.com/aeiou
        height: 1
    ArtistRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
      example:
        name: name
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
    LabelRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
      example:
        name: name
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
    Album:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        artist:
          $ref: "#/components/schemas/ArtistRef"
        label:
          $ref: "#/components/schemas/LabelRef"
        releaseDate:
          type: string
          format: date
        releaseState:
          type: string
          enum:
          - draft
          - scheduled
          - published
          - archived
        price:
          type: number
          format: float
        currency:
          type: string
          example: EUR
        genres:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        cover:
          $ref: "#/components/schemas/Image"
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/Track"
        stats:
          $ref: "#/components/schemas/Album_stats"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example:
        artist:
          name: name
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        releaseDate: 2000-01-23
        description: description
        label:
          name: name
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        title: title
        tracks:
        - createdAt: 2000-01-23T04:56:07.000+00:00
          trackNumber: 5
          artist: null
          stats:
            playCount: 7
          album:
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            title: title
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          label: null
          audio:
            codec: codec
            bitrate: 2
            url: http://example.com/aeiou
          title: title
          durationSec: 5
          lyrics:
            language: language
            text: text
          updatedAt: 2000-01-23T04:56:07.000+00:00
        - createdAt: 2000-01-23T04:56:07.000+00:00
          trackNumber: 5
          artist: null
          stats:
            playCount: 7
          album:
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            title: title
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          label: null
          audio:
            codec: codec
            bitrate: 2
            url: http://example.com/aeiou
          title: title
          durationSec: 5
          lyrics:
            language: language
            text: text
          updatedAt: 2000-01-23T04:56:07.000+00:00
        tags:
        - tags
        - tags
        cover:
          alt: alt
          width: 6
          url: http://example.com/aeiou
          height: 1
        createdAt: 2000-01-23T04:56:07.000+00:00
        stats:
          playCount: 9
          likeCount: 3
          ratingAvg: 4.5
          commentCount: 2
        price: 0.8008282
        genres:
        - genres
        - genres
        currency: EUR
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        releaseState: draft
        updatedAt: 2000-01-23T04:56:07.000+00:00
    AlbumCreateInput:
      required:
      - title
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        artistId:
          type: string
          format: uuid
        labelId:
          type: string
          format: uuid
        releaseDate:
          type: string
          format: date
        price:
          type: number
          format: float
        currency:
          type: string
          example: EUR
        genres:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
    AlbumUpdateInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        artistId:
          type: string
          format: uuid
        labelId:
          type: string
          format: uuid
        releaseDate:
          type: string
          format: date
        releaseState:
          type: string
          enum:
          - draft
          - scheduled
          - published
          - archived
        price:
          type: number
          format: float
        currency:
          type: string
          example: EUR
        genres:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
    PaginatedAlbumList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Album"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      example:
        data:
        - artist:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          releaseDate: 2000-01-23
          description: description
          label:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          title: title
          tracks:
          - createdAt: 2000-01-23T04:56:07.000+00:00
            trackNumber: 5
            artist: null
            stats:
              playCount: 7
            album:
              id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
              title: title
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            label: null
            audio:
              codec: codec
              bitrate: 2
              url: http://example.com/aeiou
            title: title
            durationSec: 5
            lyrics:
              language: language
              text: text
            updatedAt: 2000-01-23T04:56:07.000+00:00
          - createdAt: 2000-01-23T04:56:07.000+00:00
            trackNumber: 5
            artist: null
            stats:
              playCount: 7
            album:
              id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
              title: title
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            label: null
            audio:
              codec: codec
              bitrate: 2
              url: http://example.com/aeiou
            title: title
            durationSec: 5
            lyrics:
              language: language
              text: text
            updatedAt: 2000-01-23T04:56:07.000+00:00
          tags:
          - tags
          - tags
          cover:
            alt: alt
            width: 6
            url: http://example.com/aeiou
            height: 1
          createdAt: 2000-01-23T04:56:07.000+00:00
          stats:
            playCount: 9
            likeCount: 3
            ratingAvg: 4.5
            commentCount: 2
          price: 0.8008282
          genres:
          - genres
          - genres
          currency: EUR
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          releaseState: draft
          updatedAt: 2000-01-23T04:56:07.000+00:00
        - artist:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          releaseDate: 2000-01-23
          description: description
          label:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          title: title
          tracks:
          - createdAt: 2000-01-23T04:56:07.000+00:00
            trackNumber: 5
            artist: null
            stats:
              playCount: 7
            album:
              id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
              title: title
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            label: null
            audio:
              codec: codec
              bitrate: 2
              url: http://example.com/aeiou
            title: title
            durationSec: 5
            lyrics:
              language: language
              text: text
            updatedAt: 2000-01-23T04:56:07.000+00:00
          - createdAt: 2000-01-23T04:56:07.000+00:00
            trackNumber: 5
            artist: null
            stats:
              playCount: 7
            album:
              id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
              title: title
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            label: null
            audio:
              codec: codec
              bitrate: 2
              url: http://example.com/aeiou
            title: title
            durationSec: 5
            lyrics:
              language: language
              text: text
            updatedAt: 2000-01-23T04:56:07.000+00:00
          tags:
          - tags
          - tags
          cover:
            alt: alt
            width: 6
            url: http://example.com/aeiou
            height: 1
          createdAt: 2000-01-23T04:56:07.000+00:00
          stats:
            playCount: 9
            likeCount: 3
            ratingAvg: 4.5
            commentCount: 2
          price: 0.8008282
          genres:
          - genres
          - genres
          currency: EUR
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          releaseState: draft
          updatedAt: 2000-01-23T04:56:07.000+00:00
        meta:
          total: 123
          limit: 20
          page: 1
    AlbumResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Album"
      example:
        data:
          artist:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          releaseDate: 2000-01-23
          description: description
          label:
            name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          title: title
          tracks:
          - createdAt: 2000-01-23T04:56:07.000+00:00
            trackNumber: 5
            artist: null
            stats:
              playCount: 7
            album:
              id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
              title: title
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            label: null
            audio:
              codec: codec
              bitrate: 2
              url: http://example.com/aeiou
            title: title
            durationSec: 5
            lyrics:
              language: language
              text: text
            updatedAt: 2000-01-23T04:56:07.000+00:00
          - createdAt: 2000-01-23T04:56:07.000+00:00
            trackNumber: 5
            artist: null
            stats:
              playCount: 7
            album:
              id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
              title: title
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            label: null
            audio:
              codec: codec
              bitrate: 2
              url: http://example.com/aeiou
            title: title
            durationSec: 5
            lyrics:
              language: language
              text: text
            updatedAt: 2000-01-23T04:56:07.000+00:00
          tags:
          - tags
          - tags
          cover:
            alt: alt
            width: 6
            url: http://example.com/aeiou
            height: 1
          createdAt: 2000-01-23T04:56:07.000+00:00
          stats:
            playCount: 9
            likeCount: 3
            ratingAvg: 4.5
            commentCount: 2
          price: 0.8008282
          genres:
          - genres
          - genres
          currency: EUR
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          releaseState: draft
          updatedAt: 2000-01-23T04:56:07.000+00:00
    AlbumStats:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/AlbumStats_data"
      example:
        data:
          salesCount: 200
          playsLast30Days: 1200
          totalPlays: 12345
          playsLast7Days: 321
          ratingAvg: 4.5
    Track:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        durationSec:
          type: integer
        trackNumber:
          type: integer
        artist:
          $ref: "#/components/schemas/ArtistRef"
        album:
          $ref: "#/components/schemas/AlbumRef"
        label:
          $ref: "#/components/schemas/LabelRef"
        audio:
          $ref: "#/components/schemas/Track_audio"
        lyrics:
          $ref: "#/components/schemas/Track_lyrics"
        stats:
          $ref: "#/components/schemas/Track_stats"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        trackNumber: 5
        artist: null
        stats:
          playCount: 7
        album:
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          title: title
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        label: null
        audio:
          codec: codec
          bitrate: 2
          url: http://example.com/aeiou
        title: title
        durationSec: 5
        lyrics:
          language: language
          text: text
        updatedAt: 2000-01-23T04:56:07.000+00:00
    AlbumRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
      example:
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        title: title
    TrackCreateMinimal:
      required:
      - title
      type: object
      properties:
        title:
          type: string
        durationSec:
          type: integer
        artistId:
          type: string
          format: uuid
...
    Comment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        targetType:
          type: string
          enum:
          - album
          - track
          - merch
        targetId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        text:
          type: string
        rating:
          maximum: 5
          minimum: 1
          type: integer
        status:
          type: string
          default: visible
          enum:
          - visible
          - hidden
          - pending
        likes:
          type: integer
          default: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        replies:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        targetId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        replies:
        - null
        - null
        rating: 1
        targetType: album
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        text: text
        userId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        status: visible
        likes: 6
        updatedAt: 2000-01-23T04:56:07.000+00:00
    CommentCreateInput:
      required:
      - text
      type: object
      properties:
        targetType:
          type: string
          enum:
          - album
          - track
          - merch
        targetId:
          type: string
          format: uuid
        text:
          type: string
        rating:
          maximum: 5
          minimum: 1
          type: integer
    PaginatedCommentList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      example:
        data:
        - createdAt: 2000-01-23T04:56:07.000+00:00
          targetId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          replies:
          - null
          - null
          rating: 1
          targetType: album
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          text: text
          userId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          status: visible
          likes: 6
          updatedAt: 2000-01-23T04:56:07.000+00:00
        - createdAt: 2000-01-23T04:56:07.000+00:00
          targetId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          replies:
          - null
          - null
          rating: 1
          targetType: album
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          text: text
          userId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          status: visible
          likes: 6
          updatedAt: 2000-01-23T04:56:07.000+00:00
        meta:
          total: 123
          limit: 20
          page: 1
    CommentResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Comment"
      example:
        data:
          createdAt: 2000-01-23T04:56:07.000+00:00
          targetId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          replies:
          - null
          - null
          rating: 1
          targetType: album
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          text: text
          userId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          status: visible
          likes: 6
          updatedAt: 2000-01-23T04:56:07.000+00:00
    albumId_tracks_body:
      type: object
      properties:
        tracks:
          description: Array de trackIds o objetos de pista mínimos
          oneOf:
          - type: array
            items:
              type: string
              format: uuid
          - type: array
            items:
              $ref: "#/components/schemas/TrackCreateMinimal"
    albumId_cover_body:
      required:
      - file
      type: object
      properties:
        file:
          type: string
          format: binary
        alt:
          type: string
          description: Texto alternativo
...
    AlbumStats_data:
      type: object
      properties:
        totalPlays:
          type: integer
          example: 12345
        playsLast7Days:
          type: integer
          example: 321
        playsLast30Days:
          type: integer
          example: 1200
        ratingAvg:
          type: number
          format: float
          example: 4.5
        salesCount:
          type: integer
          example: 200
      example:
        salesCount: 200
        playsLast30Days: 1200
        totalPlays: 12345
        playsLast7Days: 321
        ratingAvg: 4.5
...
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.undersound.example.com/authorize
          tokenUrl: https://auth.undersound.example.com/token
          scopes:
            content.read: Lectura de contenido
            content.write: Escritura de contenido
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT